\name{bmeasures-package}
\alias{bmeasures-package}

\docType{package}
\title{
Binary similarity/dissimilarity measures
}
\description{
This package generates the quantities of the OTUs table and calculates the binary similarity/dissimilarity measures between two vectors using desired equation.
}
\details{
\tabular{ll}{
Package: \tab bmeasures\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-31\cr
License: \tab GPL-2\cr
}
\bold{Functions:}\cr
\tabular{ll}{
\code{bmeasures(x, y, method)} \tab Calculates the binary similarity/dissimilarity coefficient. \cr
\code{bmeasures_otu(x, y)} \tab Generates the quantities of the Operational Taxonomic Units table.
}
}
\author{
Sony Hartono Wijaya, Naoaki Ono, Md. Altaf-Ul-Amin, Shigehiko Kanaya\cr
Maintainer: Sony H. Wijaya <\email{sonyhartono@gmail.com}>

}
\references{
[1]  Consonni, V., & Todeschini, R. (2012). New similarity coefficients for binary data. Match-Communications in Mathematical and Computer Chemistry, 68(2), 581.

[2]  Zhang, B., & Srihari, S. N. (2001). Binary vector dissimilarity measures for handwriting identiﬁcation. In Proc. of SPIE Vol (Vol. 5010, p. 29).

[3]  Zhang, B., & Srihari, S. N. (2003, September). Properties of binary vector dissimilarity measures. In Proc. JCIS Int’l Conf. Computer Vision, Pattern Recognition, and Image Processing (Vol. 1).

[4]  Choi, S. S., Cha, S. H., & Tappert, C. C. (2010). A survey of binary similarity and distance measures. Journal of Systemics, Cybernetics and Informatics,8(1), 43-48.

[5]  Meyer, A. D. S., Garcia, A. A. F., Souza, A. P. D., & Souza Jr, C. L. D. (2004). Comparison of similarity coefficients used for cluster analysis with dominant markers in maize (Zea mays L). Genetics and Molecular Biology, 27(1), 83-91.

[6]  Warrens, M. J. (2008). Similarity coefficients for binary data: properties of coefficients, coefficient matrices, multi-way metrics and multivariate coefficients. Psychometrics and Research Methodology Group, Leiden University Institute for Psychological Research, Faculty of Social Sciences, Leiden University.

[7]	Jackson, D. A., Somers, K. M., & Harvey, H. H. (1989). Similarity coefficients: measures of co-occurrence and association or simply measures of occurrence?. American Naturalist, 436-453.

[8]	Dalirsefat, S. B., da Silva Meyer, A., & Mirhoseini, S. Z. (2009). Comparison of similarity coefficients used for cluster analysis with amplified fragment length polymorphism markers in the silkworm, Bombyx mori. Journal of Insect Science, 9.

[9]  Todeschini, R., Consonni, V., Xiang, H., Holliday, J., Buscema, M., & Willett, P. (2012). Similarity coefficients for binary chemoinformatics data: overview and extended comparison using simulated and real data sets. Journal of chemical information and modeling, 52(11), 2884-2901.

[10]	Jaccard, P. (1912). The distribution of the flora in the alpine zone. 1. New phytologist, 11(2), 37-50.

[11]	Dice, L. R. (1945). Measures of the amount of ecologic association between species. Ecology, 26(3), 297-302.

[12]	Hubalek, Z. (1982). Coefficients of association and similarity, based on binary (presence-absence) data: an evaluation. Biological Reviews, 57(4), 669-689.

[13]	Cheetham, A. H., & Hazel, J. E. (1969). Binary (presence-absence) similarity coefficients. Journal of Paleontology, 1130-1136.

[14]	Cha, S. H., Choi, S., & Tappert, C. C. (2009). Anomaly between jaccard and tanimoto coefficients. Proceedings of Student-Faculty Research Day, CSIS, Pace university.

[15]	Cha, S. H., Tappert, C., & Yoon, S. (2006). Enhancing Binary Feature Vector Similarity Measures. Journal of Pattern Recognition Research, 1, 63-77.

[16]	Lourenco, F., Lobo, V., & Bacao, F. (2004). Binary-based similarity measures for categorical data and their application in Self-Organizing Maps.

[17]	Warrens, M. J. (2008). Similarity coefficients for binary data: properties of coefficients, coefficient matrices, multi-way metrics and multivariate coefficients. Psychometrics and Research Methodology Group, Leiden University Institute for Psychological Research, Faculty of Social Sciences, Leiden University.

[18]	Ojurongbe, T. A. (2012). Comparison of different proximity measures and classification methods for binary data.

[19]	Johnson, S. C. (1967). Hierarchical clustering schemes. Psychometrika, 32(3), 241-254.

[20]	Michael, E. L. (1920). Marine ecology and the coefficient of association: a plea in behalf of quantitative biology. Journal of Ecology, 8(1), 54-59.

[21]	Stiles, H. E. (1961). The association factor in information retrieval. Journal of the ACM (JACM), 8(2), 271-279.

[22]	Nei, M., & Li, W. H. (1979). Mathematical model for studying genetic variation in terms of restriction endonucleases. Proceedings of the National Academy of Sciences, 76(10), 5269-5273.

[23]	Holliday, J. D., Hu, C. Y., & Willett, P. (2002). Grouping of coefficients for the calculation of inter-molecular similarity and dissimilarity using 2D fragment bit-strings. Combinatorial chemistry & high throughput screening, 5(2), 155-166.
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
%% \seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%% }
%% \examples{
%% 
%% }
