\name{bmeasures-package}
\alias{bmeasures-package}

\docType{package}
\title{
Binary similarity/dissimilarity measures
}
\description{
This package generates the quantities of the OTUs table and calculates the binary similarity/dissimilarity measures between two vectors using desired equation.
}
\details{
\tabular{ll}{
Package: \tab bmeasures\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-31\cr
License: \tab GPL-2\cr
}
\bold{Functions:}\cr
\tabular{ll}{
\code{bmeasures(x, y, method)} \tab Calculates the binary similarity/dissimilarity coefficient. \cr
\code{bmeasures_otu(x, y)} \tab Generates the quantities of the Operational Taxonomic Units table.
}
}
\author{
Sony Hartono Wijaya, Naoaki Ono, Md. Altaf-Ul-Amin, Shigehiko Kanaya\cr
Maintainer: Sony H. Wijaya <\email{sonyhartono@gmail.com}>

}
\references{
[1]  Choi, S. S., Cha, S. H., & Tappert, C. C. (2010). A survey of binary similarity and distance measures. Journal of Systemics, Cybernetics and Informatics,8(1), 43-48.

[2]  Consonni, V., & Todeschini, R. (2012). New similarity coefficients for binary data. Match-Communications in Mathematical and Computer Chemistry, 68(2), 581.\cr

[3]  Zhang, B., & Srihari, S. N. (2001). Binary vector dissimilarity measures for handwriting identiﬁcation. In Proc. of SPIE Vol (Vol. 5010, p. 29).

[4]  Zhang, B., & Srihari, S. N. (2003, September). Properties of binary vector dissimilarity measures. In Proc. JCIS Int’l Conf. Computer Vision, Pattern Recognition, and Image Processing (Vol. 1).

[5]  Warrens, M. J. (2008). Similarity coefficients for binary data: properties of coefficients, coefficient matrices, multi-way metrics and multivariate coefficients. Psychometrics and Research Methodology Group, Leiden University Institute for Psychological Research, Faculty of Social Sciences, Leiden University.

[6]	Jackson, D. A., Somers, K. M., & Harvey, H. H. (1989). Similarity coefficients: measures of co-occurrence and association or simply measures of occurrence?. American Naturalist, 436-453.

[7]	Dalirsefat, S. B., da Silva Meyer, A., & Mirhoseini, S. Z. (2009). Comparison of similarity coefficients used for cluster analysis with amplified fragment length polymorphism markers in the silkworm, Bombyx mori. Journal of Insect Science, 9.
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
%% \seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%% }
%% \examples{
%% 
%% }
